requirements.txt
llama-index
llama-parse
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-vector-stores-chroma
chromadb
python-dotenv
pandas
jupyter
pydantic
